examples/financial-services-advisor/google-cloud-financial-advisor/backend/pyproject.toml
[project]
name = "google-cloud-financial-advisor"
version = "0.1.0"
description = "AI-powered financial compliance assistant using Google ADK and Neo4j Agent Memory"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Neo4j Labs" }
]
keywords = ["neo4j", "agent-memory", "google-adk", "financial-services", "compliance", "aml", "kyc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.8.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "google-adk>=0.3.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-genai>=0.3.0",
    "neo4j-agent-memory[google-adk,vertex-ai]>=0.1.0",
    "neo4j>=5.14.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true

[tool.uv.sources]
# For local development, use the editable path:
neo4j-agent-memory = { path = "../../../..", editable = true }
# For deployment (production), use the Git URL:
# neo4j-agent-memory = { git = "https://github.com/neo4j-labs/agent-memory" }

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
